quinta sesión - vistas, plantillas, javascript, css y arte (ficheros estáticos)

+++ apps/vemosla_web/lib/vemosla_web/controllers/post_controller.ex
@@ -17,22 +17,30 @@ defmodule VemoslaWeb.PostController do
     render(conn, "new.html", changeset: changeset)
   end
 
-  defp reactions(%{"reactions" => %{"0" => %{"watched" => "true"}}} = params) do
-    put_in(
-      params["reactions"]["0"]["reaction"],
+  defp reactions(%{"reactions" => %{"0" => %{"watched" => "true"}}} = params, user) do
+    params
+    |> put_in(
+      ["reactions", "0", "reaction"],
       params["reactions"]["0"]["reaction_watched"]
     )
+    |> put_in(["reactions", "0", "user_id"], user.id)
   end
-  defp reactions(%{"reactions" => %{"0" => %{"watched" => "false"}}} = params) do
-    put_in(
-      params["reactions"]["0"]["reaction"],
+
+  defp reactions(%{"reactions" => %{"0" => %{"watched" => "false"}}} = params, user) do
+    params
+    |> put_in(
+      ["reactions", "0", "reaction"],
       params["reactions"]["0"]["reaction_non_watched"]
     )
+    |> put_in(["reactions", "0", "user_id"], user.id)
   end
 
   def create(conn, %{"post" => post_params}) do
+    user = conn.assigns.current_user
+
     post_params
-    |> reactions()
+    |> Map.put("user_id", user.id)
+    |> reactions(user)
     |> Timeline.create_post()
     |> case do
       {:ok, _post} ->
@@ -46,11 +54,19 @@ defmodule VemoslaWeb.PostController do
   end
 
   def delete(conn, %{"id" => id}) do
+    user = conn.assigns.current_user
     post = Timeline.get_post!(id)
-    {:ok, _post} = Timeline.delete_post(post)
 
-    conn
-    |> put_flash(:info, "Post deleted successfully.")
-    |> redirect(to: Routes.post_path(conn, :index))
+    if post.user_id == user.id do
+      {:ok, _post} = Timeline.delete_post(post)
+
+      conn
+      |> put_flash(:info, "Publicación eliminada.")
+      |> redirect(to: Routes.post_path(conn, :timeline))
+    else
+      conn
+      |> put_flash(:error, "No puedes borrar una publicación que no es tuya.")
+      |> redirect(to: Routes.post_path(conn, :timeline))
+    end
   end
 end
